Issue #SB-8131 fix: Remove profile from header menu and add workspace to header menu
