+++ src/app/resourcebundles/data/formElements/en.properties
@@ -449,6 +449,7 @@ courses=Courses
 resources=Library
 community=Groups
 profile=Profile
+workspace=Workspace
